+++ worker.py
@@ -1,5 +1,6 @@
 import os
-from rq import Worker, Queue, Connection
+from rq import Queue
+from rq.worker import Worker
 import redis
 
 listen = ['default']
@@ -7,6 +8,6 @@ redis_url = os.getenv('REDIS_URL', 'redis://localhost:6379/0')
 conn = redis.from_url(redis_url)
 
 if __name__ == '__main__':
-    with Connection(conn):
-        worker = Worker(list(map(Queue, listen)))
-        worker.work() 
+    queues = [Queue(name, connection=conn) for name in listen]
+    worker = Worker(queues, connection=conn)
+    worker.work()
